requirements.txt
cffi==2.0.0
cryptography==46.0.4
pillow==12.1.0
pycparser==3.0
